+++ src/_types/index.d.ts
@@ -47,7 +47,7 @@ export interface auth0__password_confirmation__error_T extends Auth0Error {
 export declare type change_password_error_I = auth0__password_confirmation__error_T
 export interface auth0__userinfo__GET__fetch2__params_T {
 	auth0__token:JwtToken
-	AUTH0_DOMAIN:string
+	AUTH0_DOMAIN?:string
 }
 export type auth0__userinfo__fetch_get__params_T = auth0__userinfo__GET__fetch2__params_T
 export type get_auth0_userinfo_params_I = auth0__userinfo__GET__fetch2__params_T